Move HotspotInfo to it's own directory add index.js add css file
