+++ src/HotspotInfo/HotspotInfo.js
