+++ src/HotspotInfo/index.js
@@ -1,0 +1,3 @@
+import { HotspotInfo } from './HotspotInfo';
+
+export default HotspotInfo;
